Initial commit
This is a simple ticker that pulls in the price of a Bitcoin and
displays the price in a variety of currencies.
The prices are pulled from either Bitstamp or BTC-e, and are updated
every 10 seconds. Exchange rates are pulled from
www.getexchangerates.com. Vector graphics were pulled from a forgotten
public resource, but are much appreciated.